requirements.txt
email-validator>=2.2.0
flask>=3.1.0
flask-sqlalchemy>=3.1.1
google-cloud-vision>=3.10.1
google-generativeai>=0.8.5
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
requests>=2.32.3
sift-stack-py>=0.5.1
werkzeug>=3.1.3
python-dotenv>=1.0.0
Flask-Migrate>=4.0.4
Pillow>=9.0.0
pytesseract>=0.3.10

# RAG & embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.25.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
